Phase 2: Migrate main/services/ to TypeScript
- Restored 12 existing .ts service files from git
- Fixed TypeScript errors in keyword-enrichment.ts and SubjectService.ts
- Used proper Date.getTime() for arithmetic operations
- Fixed function signatures to match expected parameter counts
- Used Set.has() instead of Set.includes()
- Replaced calculateIdHashOfObj with crypto.createHash for signature generation
- All service files now compile without errors
